algoritmo guloso metade implementado
