--- Greedy.cpp
@@ -1,3 +1,0 @@
-#include "Greedy.h"
-
-
