--- Greedy.h
@@ -1,14 +1,0 @@
-#ifndef GREEDY_
-#define GREEDY_
-
-#include "JobXServer.h"
-
-
-
-class Greedy{
-    public:
-        
-
-};
-
-#endif
