--- Jobs.h
@@ -1,14 +1,0 @@
-#ifndef JOBS_H
-#define JOBS_H
-
-class Jobs{
-    public:
-        int id, tempo;
-
-        Jobs(int id, int tempo){
-            this->id = id;
-            this->tempo = tempo;
-        }
-};
-
-#endif
