+++ classes/func/Jobs.cpp
@@ -1,0 +1,7 @@
+#include "../interfaces/Jobs.h"
+
+Jobs::Jobs(int id, int tempo, int custo){
+    this->id = id;
+    this->tempo = tempo;
+    this->custo = custo;
+}
