+++ classes/interfaces/Jobs.h
@@ -1,0 +1,10 @@
+#ifndef JOBS_H
+#define JOBS_H
+
+class Jobs{
+    public:
+        int id, tempo, custo;
+        Jobs(int id, int tempo, int custo);
+};
+
+#endif
